requirements.txt
numpy
mediapipe
opencv-python
Pillow
diffusers
transformers
accelerate